Add auto test for chr20, split files to make them testable
